moved functions and classes to utils
